feat(validators)!: Change to more simple interface
Didn't feel like it made sense for the interface to return *some*
value related to the validation. Ultimately, it should be usable
to just validate and see whether the content is valid. Some
related value can be otherwise added if necessary, but is less
relevant and makes the interface less useful if being directly passed
as a generic type.