feat(validators)!: Simplify HeaderValidator
Realised that not everybody might care about the verb, as it's also
not a part of Conventional commits, so decided to make it optional.
In doing so, noticed that the same could be done with the other
slices and parts, making it easier have the "empty" validator still be useful,
though it does still need to be initialised due to the regexp. Technically
not necessary as the regexp is the same regardless,
but I do feel like may be better to just keep it as having a New* function,
in case there's something added that couldn't be done using the defaults.