feat!: Add parsing validators from YAML
Makes it possible to have a config file for configuring
the validator.

This is a rebase commit, the rest are constituent commits.

BREAKING-CHANGE: HeaderValidator's SetHeaderLength renamed to SetLineLength.
BREAKING-CHANGE: HeaderValidator's ValidateHeaderLength renamed to
  ValidateLength.
BREAKING-CHANGE: TrailerValidator's key-value validation no longer checks
  the length of the line; the line length is checked separately during overall
  validation.
BREAKING-CHANGE: TrailerValidator's constructor will no longer panic if its
  regex isn't successfully compiled.
BREAKING-CHANGE: Child validators in MessageValidator changed to pointers.
Effect: behavioural, API

feat(validators): Add YAML unmarshal for header

Not entirely sure how to construct it, really. Ultimately, each of the options
should be a part of the whole YAML file, so the header would be under its own
key. At the same time, I don't necessarily want to hardcode that it needs to
pick up a "header.types" key or whatever, because it could also be something
like "validator.header.types" or anything else. Could also define the entire
file's expected structure as the thing that gets unmarshaled (maybe? does it
unmarshal nested structs?), but I don't know how usable this is.

Ideally, the
individual parts should be each defineable irrespective of each other and the
final format of the config file as much as possible. so that it's possible to
cleanly update one part without having to consider the others -- basic stuff.
Feeling like maybe the best approach is to simply pick individual parts from
the file using something like path https://pkg.go.dev/github.com/goccy/go-yaml@v1.19.1#Path,
passing that info somehow to the unmarshaling methods? I'd see that using the
oen big struct constructed out of the smaller ones should allow the parsing to
happen in one pass, where as using Paths may not allow this quite as cleanly,
so not sure.

At the end of the day, though, it's not like it's going to be prohibitively
slow to just use Paths, so probably will just use those.

feat(validators): Add YAML unmarshal for body

feat(validators): Add YAML unmarshal for trailer

feat(validators): Add Equal method to header

Can't compare directly because of the regexp, so added this. It should match
the go-cmp package as well, so if that's ever used, this is ready to go. Needed
for the purposes of testing, but doesn't hurt to have in general, of course.

feat(validators)!: Change naming for header validator to be more consistent

The use of lineLength to represent the length of the header line is more in line
with the body and trailer validators' naming schemes.

BREAKING-CHANGE: HeaderValidator's SetHeaderLength renamed to SetLineLength.
BREAKING-CHANGE: HeaderValidator's ValidateHeaderLength renamed to
  ValidateLength.
Effect: API

feat(validators)!: Change trailer's key-value validation

Having the line length be validated in the key-value pair validation isn't
really the expected behaviour to begin with, and was causing problems with
the flow of the validation. Now validated separately.

BREAKING-CHANGE: TrailerValidator's key-value validation no longer checks
  the length of the line; the line length is checked separately during overall
  validation.
Effect: behavioural

feat(validators)!: Change trailer's constructor

Adding a continuationIndent setter for testing purposes; using just Compile
for the regex because panicking isn't how the rest of everything works
(returning errors as a result of failure). Panic is only sensible for global,
constant regex compilation.

Also changing the regex to pointer because that's how it is in the header.
Not entirely necessary, perhaps, as with the header the regex is a global
one, so space is saved, whereas here the regex is technically unique to
each TrailerValidator, so you're creating a new regex object each time
regardless. It just doesn't directly make the size of the validator bigger.
Saving space isn't a major concern anyway, really, but whatever.

BREAKING-CHANGE: TrailerValidator's constructor will no longer panic if its
  regex isn't successfully compiled.
Effect: behavioural

feat(validators): Add YAML testing for MessageValidator

No implementation yet.

fix(validators): Fix setting of continuation indent

Wasn't setting the regex when setting the continuation indent.

Effect: behavioural

feat(validators): Add YAML unmarshal for message

Effect: behavioural

feat(validators)!: Change to using pointers in MessageValidators

Makes some stuff a little easier, and makes sense to use pointers.

BREAKING-CHANGE: Child validators in MessageValidator changed to pointers.

fix(validators): Fix checking of required trailer keys in message

The keys were not being checked if only a header was found.

Effect: behavioural